+++ package.json
@@ -14,6 +14,7 @@
     "electron-rebuild": "^1.5.7",
     "lodash": "^4.17.4",
     "moment": "^2.18.1",
+    "octonode": "^0.7.11",
     "react": "^15.4.2",
     "react-collapse": "^3.3.0",
     "react-dom": "^15.4.2",
